add a wrapper for jenkins pull request build
